+++ ft_putstr.c
@@ -6,14 +6,14 @@
 /*   By: mirsella <mirsella@protonmail.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 16:39:49 by mirsella          #+#    #+#             */
-/*   Updated: 2022/11/21 16:43:38 by mirsella         ###   ########.fr       */
+/*   Updated: 2022/11/21 17:13:13 by mirsella         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "unistd.h"
 #include "libft.h"
 
-size_t	ft_putstr(char *str)
+int	ft_putstr(char *str)
 {
 	int	i;
